# swagger.yml
openapi: "3.0.3"
info:
  title: "ACE Telemetry API"
  description: "Documentation for Ace Telemetry API. This telemetry server implements a longitudinal tracking system   for ACE Data Bottles.  When `ace-dt bottle push` or `ace-dt bottle pull` is run it pushed the bottle (metadata and data) to the OCI registry that is provided.  In addition those commands also push the metadata to 0 or more telemetry servers configured in ace-dt's configuration file.  In this way the telemetry server can be used to discover and track data bottles."
  version: "0.8.19"
  contact:
    email: "Kyle.Tarplee@udri.udayton.edu"
...
paths:
...
  /api/metric:
    get:
      tags:
        - "metric"
      summary: "This will Get Bottle details using metric data"
      operationId: "handleGetBottlesFromMetric"
      parameters:
        - name: "selector"
          in: "query"
          description: "selector parameter"
          required: false
          schema:
            type: "string"
        - name: "metric"
          in: "query"
          description: "Description parameter"
          required: false
          schema:
            type: "string"
        - name: "limit"
          in: "query"
          description: "limit parameter"
          required: false
          schema:
            type: "integer"
        - name: "digestOnly"
          in: "query"
          description: "DigestOnly parameter"
          required: false
          schema:
            type: "boolean"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                items:
                  $ref: "#/components/schemas/ResponseObj"
        "400":
          description: "Invalid status value"
  /api/location:
    get:
      tags:
        - "location"
      summary: "This gets the Location for the bottle using its bottle_digest parameter"
      operationId: "handleGetLocation"
      parameters:
        - name: "bottle_digest"
          in: "query"
          description: "{bottle digest} param"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                items:
                  $ref: "#/components/schemas/ResponseObj"
        "400":
          description: "Invalid status value"
          #  ---- /Represent Blob/Artifact Information  ----------------------------------------
...
servers:
  - url: 'http://localhost:8100'
    description: Local dev server
#  ---- /Definitions below  ----------------------------------------
components:
  schemas:
    ResponseObj:
      type: "object"
      properties:
        HTTP/1.1:
          type: "integer"
        Content-Type:
          type: "string"
        Date:
          type: "string"
        Content-Length:
          type: "integer"
        Connection:
          type: "string"
...